create markup file
