+++ src/js/reviews-api.js
@@ -4,9 +4,7 @@ import iziToast from 'izitoast';
 
 import 'izitoast/dist/css/iziToast.min.css';
 
-const swiperList = document.querySelector('.swiper-wrapper');
-
-const buttonWrapper = document.querySelector('.button-wrapper');
+import { createMarkup, notFound } from './reviews-markup';
 
 async function fetchReviews(params) {
   try {
@@ -22,31 +20,6 @@ async function fetchReviews(params) {
       message: error.message,
     });
   }
-}
-
-function createMarkup(data) {
-  const markup = data
-    .map(({ author, avatar_url, review }) => {
-      return `<li class="swiper-slide">
-					<p class="reviews-text">${review}</p>
-					<div class="card-wrapper">
-						<img class="reviews-icon" src="${avatar_url}" alt="${author}">
-						<h2 class="reviews-author">${author}</h2>
-					</div>
-				</li>`;
-    })
-    .join('');
-  swiperList.innerHTML = markup;
-  buttonWrapper.classList.remove('visually-hidden');
-}
-
-function notFound() {
-  const notFoundMarkup = `<li class="not-found swiper-slide">
-	<p class="not-found-text">Not found!</p>
-</li>`;
-  swiperList.innerHTML = notFoundMarkup;
-
-  //   buttonWrapper.classList.remove('visually-hidden');
 }
 
 fetchReviews();
