+++ src/js/reviews-swiper.js
@@ -23,4 +23,34 @@ const swiper = new Swiper('.swiper', {
 
   slidesPerView: 1,
   spaceBetween: 32,
+
+  //   breakpoints: {
+  //     320: {
+  //       slidesPerView: 2,
+  //       spaceBetween: 32,
+  //     },
+  //     768: {
+  //       slidesPerView: 3,
+  //       spaceBetween: 32,
+  //     },
+  //     1280: {
+  //       slidesPerView: 4,
+  //       spaceBetween: 32,
+  //     },
+  //   },
+
+  //   breakpoints: {
+  //     640: {
+  //       slidesPerView: 2,
+  //       spaceBetween: 20,
+  //     },
+  //     768: {
+  //       slidesPerView: 4,
+  //       spaceBetween: 40,
+  //     },
+  //     1024: {
+  //       slidesPerView: 5,
+  //       spaceBetween: 50,
+  //     },
+  //   },
 });
