+++ src/partials/reviews-section.html
@@ -10,10 +10,10 @@
 
 		<div class="button-wrapper visually-hidden">
 			<button class="swiper-button-prev"><svg class="" width="20px" height="20px">
-					<use href="../img/icon.svg#icon-arrow-left2"></use>
+					<use href="/img/icon.svg#icon-arrow-left2"></use>
 				</svg></button>
 			<button class="swiper-button-next"><svg class="icon-svg" width="20px" height="20px">
-					<use href="../img/icon.svg#icon-arrow-right2"></use>
+					<use href="/img/icon.svg#icon-arrow-right2"></use>
 				</svg></button>
 		</div>
 	</div>
